feat: issue-105 - Step 1: Remove MongoDB runtime dependency and repository code
