+++ legacy/product-service/src/main/java/pl/altkom/asc/lab/micronaut/poc/product/service/infrastructure/adapters/db/ProductsRepository.java
@@ -1,52 +1,15 @@
 package pl.altkom.asc.lab.micronaut.poc.product.service.infrastructure.adapters.db;
 
-import com.mongodb.client.model.Filters;
-import com.mongodb.reactivestreams.client.MongoClient;
-import com.mongodb.reactivestreams.client.MongoCollection;
-import io.reactivex.Flowable;
-import io.reactivex.Maybe;
-import io.reactivex.Single;
-import io.micronaut.context.annotation.Requires;
-import lombok.RequiredArgsConstructor;
-import pl.altkom.asc.lab.micronaut.poc.product.service.domain.Product;
-import pl.altkom.asc.lab.micronaut.poc.product.service.domain.Products;
+import io.micronaut.data.annotation.Repository;
+import io.micronaut.data.repository.CrudRepository;
+import pl.altkom.asc.lab.micronaut.poc.product.service.infrastructure.adapters.db.model.ProductEntity;
 
-import javax.inject.Singleton;
-import java.util.List;
+import java.util.Optional;
 
-@Singleton
-@RequiredArgsConstructor
-@Requires(property = "products.persistence", value = "mongo")
-public class ProductsRepository implements Products {
+@Repository
+public interface ProductsRepository extends CrudRepository<ProductEntity, String> {
 
-    private final MongoClient mongoClient;
+    Optional<ProductEntity> findByCode(String code);
 
-    @Override
-    public Single<Product> add(Product product) {
-        return Single.fromPublisher(
-                getCollection().insertOne(product)
-        ).map(success -> product);
-    }
-
-    @Override
-    public Single<List<Product>> findAll() {
-        return Flowable.fromPublisher(
-                getCollection().find()
-        ).toList();
-    }
-
-    @Override
-    public Maybe<Product> findOne(String productCode) {
-        return Flowable.fromPublisher(
-                getCollection()
-                        .find(Filters.eq("code", productCode))
-                        .limit(1)
-        ).firstElement();
-    }
-
-    private MongoCollection<Product> getCollection() {
-        return mongoClient
-                .getDatabase("products-demo")
-                .getCollection("product", Product.class);
-    }
+    ProductEntity save(ProductEntity productEntity);
 }
